+++ src/src3/Helpers/LexerBase.hpp
@@ -13,9 +13,9 @@ size_t p_column = 1;
 size_t p_line = 1;
 
 inline void reset() {p_index = 0; p_line = 1; p_column = 1;}
-char peek(size_t distance = 0) const;
+char peek(int distance = 0) const;
 char consume();
-bool match(char c, size_t distance = 0) const;
+bool match(char c, int distance = 0) const;
 inline bool notAtEnd() const {return p_index < p_source.size();}
 inline bool isAtEnd() const {return !(p_index < p_source.size());}
